django_proj/dj-ninja-pi/pyproject.toml
[tool.poetry]
name = "dj-ninja-pi"
version = "0.1.0"
description = ""
authors = ["Vinceeee <491323832@qq.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "<5"
django-ninja = "^1.1.0"
uvicorn = "^0.30.1"
channels = "^4.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-django = "^4.7.0"
ipdb = "^0.13.13"
pytest-cov = "^4.1.0"


[[tool.poetry.source]]
name = "tencent"
url = "https://mirrors.tencent.com/pypi/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
